adds new pics to gallery and contact

+++ src/components/header/Header.jsx
@@ -16,14 +16,16 @@ const Header = ({ Link, NavLink }) => {
     useEffect(()=> {
         if (message === null) {
             API.getImportantNews(res => {
-                if(res && res.length) {
-                    res.forEach(content => {
-                        if (moment().format(content.endDate) > moment().format()) {
-                            setMessage(content.message);
-                        } else {
-                            API.deleteImportantNews(content.id, res => {});
-                        }
-                    })
+                if (res !== 'error') {
+                    if(res && res.length) {
+                        res.forEach(content => {
+                            if (moment().format(content.endDate) > moment().format()) {
+                                setMessage(content.message);
+                            } else {
+                                API.deleteImportantNews(content.id, res => {});
+                            }
+                        })
+                    }
                 }
             });
         }
